+++ client/src/createUser.js
@@ -1,27 +1,33 @@
 import hashPassword from './hashPassword'
 import gql from 'graphql-tag'
-import {storeLoginToken} from './store'
+import {handleLoginCallback, getClient, startLoggingIn, endLoggingIn} from './store'
 
-export default async function ({username, email, password, profile}, apollo) {
-  const result = await apollo.mutate({
-    mutation: gql`
-    mutation createUser ($username: String, $email: String, $password: HashedPassword!, $profile: CreateUserProfileInput) {
-      createUser (username: $username, email: $email, password: $password, profile: $profile) {
-        id
-        token
-        tokenExpires
+export default async function ({username, email, password, profile}) {
+  startLoggingIn()
+  let result
+  try {
+    result = await getClient().mutate({
+      mutation: gql`
+      mutation createUser ($username: String, $email: String, $password: HashedPassword!, $profile: CreateUserProfileInput) {
+        createUser (username: $username, email: $email, password: $password, profile: $profile) {
+          id
+          token
+          tokenExpires
+        }
       }
-    }
-    `,
-    variables: {
-      username,
-      email,
-      password: hashPassword(password),
-      profile
-    }
-  })
+      `,
+      variables: {
+        username,
+        email,
+        password: hashPassword(password),
+        profile
+      }
+    })
+  } catch (err) {
+    return handleLoginCallback(err)
+  } finally {
+    endLoggingIn()
+  }
 
-  const {id, token, tokenExpires} = result.data.createUser
-  await storeLoginToken(id, token, new Date(tokenExpires))
-  return id
+  return handleLoginCallback(null, result.data.createUser)
 }