+++ client/src/oauth/loginWithFacebook.js
@@ -1,28 +1,34 @@
 import gql from 'graphql-tag'
-import {storeLoginToken} from '../store'
+import {handleLoginCallback, getClient, startLoggingIn, endLoggingIn} from '../store'
 
 /**
  * Pass the accessToken
  * It's recommended to use https://github.com/keppelen/react-facebook-login
  */
 
-export default async function ({accessToken}, apollo) {
-  const result = await apollo.mutate({
-    mutation: gql`
-    mutation loginWithFacebook ($accessToken: String!) {
-      loginWithFacebook (accessToken: $accessToken) {
-        id
-        token
-        tokenExpires
+export default async function ({accessToken}) {
+  startLoggingIn()
+  let result
+  try {
+    result = await getClient().mutate({
+      mutation: gql`
+      mutation loginWithFacebook ($accessToken: String!) {
+        loginWithFacebook (accessToken: $accessToken) {
+          id
+          token
+          tokenExpires
+        }
       }
-    }
-    `,
-    variables: {
-      accessToken
-    }
-  })
+      `,
+      variables: {
+        accessToken
+      }
+    })
+  } catch (err) {
+    return handleLoginCallback(err)
+  } finally {
+    endLoggingIn()
+  }
 
-  const {id, token, tokenExpires} = result.data.loginWithFacebook
-  await storeLoginToken(id, token, new Date(tokenExpires))
-  return id
+  return handleLoginCallback(null, result.data.loginWithFacebook)
 }